code/requirements.txt
matplotlib==3.6.1
numpy==1.23.4
Pillow==9.3.0
scipy==1.9.3
tensorboard==2.11.0
torch==1.12.1
torchvision==0.13.0a0
tqdm==4.64.1
